# docs/swagger.yaml
basePath: /
definitions:
  app.LoginReqBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  app.SignUpReqBody:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  app.errorResponse:
    properties:
      message:
        type: string
    type: object
  ds.Collaboration:
    properties:
      id:
        type: string
      project_id:
        type: string
      user_id:
        type: string
    type: object
  ds.FavoriteProject:
    properties:
      id:
        type: string
      project_id:
        type: string
      user_id:
        type: string
    type: object
  ds.Notification:
    properties:
      deadline:
        type: string
      description:
        type: string
      error_status:
        type: integer
      id:
        type: string
      section_id:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  ds.Project:
    properties:
      color:
        type: string
      description:
        type: string
      id:
        type: string
      last_edited:
        type: string
      owner_id:
        type: string
      title:
        type: string
    required:
    - color
    - description
    - title
    type: object
  ds.Section:
    properties:
      color:
        type: string
      id:
        type: string
      project_id:
        type: string
      title:
        type: string
    type: object
  ds.User:
    properties:
      email:
        type: string
      id:
        type: string
      is_manager:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: mkn-notifyer@mail.ruv
    name: MKN Support
  description: Notification backend service.
  title: MKN API
  version: "1.0"
paths:
  /email:
    put:
      description: Changes user email
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.FavoriteProject'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Changes user email
      tags:
      - change
  /favorite:
    post:
      description: Add favorite project
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.FavoriteProject'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      security:
      - BearerAuth: []
      summary: Add favorite project
      tags:
      - add
  /favorite/{project_id}:
    delete:
      description: Delete favorite user project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.FavoriteProject'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete favorite project
      tags:
      - delete
  /favorites:
    get:
      description: Returns favorite projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.FavoriteProject'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets favorite projects
      tags:
      - info
  /favorites/{project_id}:
    get:
      description: Returns favorite projects
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.FavoriteProject'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets favorite projects
      tags:
      - info
  /login:
    post:
      description: Login user
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.LoginReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Login user
      tags:
      - auth
  /logout:
    get:
      description: Logout user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.User'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /project:
    post:
      description: Creates project
      parameters:
      - description: Project data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/ds.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Project'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      security:
      - BearerAuth: []
      summary: Creates project
      tags:
      - add
  /project/{project_id}:
    delete:
      description: Deletes a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Project'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Deletes project
      tags:
      - delete
    put:
      description: Updates a specific project according to the entered parameters
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Project'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Update project
      tags:
      - change
  /project/{project_id}/collaborator:
    delete:
      description: Removes a collaborator from the current project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Collaboration'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Deletes collaborator
      tags:
      - delete
    post:
      description: Adds a collaborator to the current project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Collaboration'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Adds collaborators
      tags:
      - add
  /project/{project_id}/collaborators:
    get:
      description: Returns all collaborators of the project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Collaboration'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Returns collaborators
      tags:
      - info
  /project/{project_id}/section:
    post:
      description: Creates a section in the project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Section'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Creates section
      tags:
      - add
  /project/{project_id}/sections:
    get:
      description: Returns all sections of the current project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Section'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Returns all sections
      tags:
      - info
  /project/section/{section_id}:
    delete:
      description: Deletes section from current project
      parameters:
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Section'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Deletes section
      tags:
      - delete
    put:
      description: Updates a section in the current project
      parameters:
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Section'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Updates section
      tags:
      - change
  /project/section/{section_id}/notification:
    post:
      description: Creates notification in accordance with the entered parameters
      parameters:
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Notification'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Creates notification
      tags:
      - add
  /project/section/{section_id}/notifications:
    get:
      description: Returns all notifications in the current section
      parameters:
      - description: Section ID
        in: path
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Notification'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets All Notifications
      tags:
      - info
  /project/section/notification/{notification_id}:
    delete:
      description: Update information about a specific notification
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Notification'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Delete notifications
      tags:
      - delete
    get:
      description: Returns Notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Notification'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets Notification
      tags:
      - info
    put:
      description: Update information about a specific notification according to the
        entered parameters
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Notification'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Update notifications
      tags:
      - change
  /projects:
    get:
      description: Returns all projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Project'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets all projects
      tags:
      - info
  /projects/latest:
    get:
      description: Returns the last three projects by last edit time
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Project'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Returns last 3 projects
      tags:
      - info
  /projects/owned:
    get:
      description: Gets all owned project
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Project'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      security:
      - BearerAuth: []
      summary: Gets all owned project
      tags:
      - info
  /signup:
    post:
      description: Signup user
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.SignUpReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Signup user
      tags:
      - auth
  /undelivered_notifications:
    get:
      description: Allows the manager to view undelivered notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Notification'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets undelivired notifications
      tags:
      - info
  /upcoming:
    get:
      description: Returns upcoming notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Notification'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.errorResponse'
      summary: Gets upcoming notifications
      tags:
      - info
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"